Archive of RProject, R Script and Input data set
This Archive contains below:

Your Task
You have been asked by Danielle Sherman, CTO of Blackwell Electronics, to predict the customers' brand preferences that are missing from the incomplete surveys by conducting two classification methods in R. Once you have determined which classifier — C5.0 or RandomForest —works better on the provided data set, she would like you to predict the brand preferences for the incomplete survey responses and prepare a report of your findings.

This task requires you to prepare one deliverable for Danielle Sherman:

Customer Brand Preferences Report. A report in a Zip file that includes:
A brief summary in Word or Excel of your methods and results that includes:
The classifiers you tried.
The classifier you selected to make the predictions, including a rationale for selecting the method you did and the level of confidence in the predictions.
The predicted answers to the brand preference question for the instances of survey results that are missing that answer.
A chart that displays the customer preference for each brand based on the combination of the actual answers and the predicted answers to the brand preference survey question.
The results of each classifier you ran exported from R
The steps in the following tabs will walk you through this process.


*** Plan of attack
Import and familiarize yourself with the training set. The file CompleteResponses is an CSV file that includes answers to all the questions in the market research survey, including which brand of computer products the customer prefers. You will be using this data set to train and test the classifier to make predictions. You will notice that there is a mix of numeric and nominal values; some might need to be converted to factors. Consult the Survey Key to gain an understanding of what the survey response values mean. The coding key’s primary function is to help you understand the coding of the survey – you only need the coding key to prepare the labels in your final preferences graph. 
Perform any necessary pre-processing.
Perform any necessary Feature Engineering/Selection
Using createDataPartition create training and test sets. These will be created from the CompleteResponses.csv file. The training data should represent 75% of the total data and the remaining 25% will be used for testing. After optimizing your model you'll later use it to make predictions on the incomplete surveys.

Predictions:
Build a model using a decision tree, C5.0,  on the training set with 10-fold cross validation and an Automatic Tuning Grid. In an earlier task, you used another decision trees, Gradient Boosted Trees, to predict a specific numeric value (the sales volume of a new product); in this task you will use C5.0 to predict nominal data (a customer's computer brand preference). The main difference is that, in numeric prediction, you are inferring a real number (such as how many products may be sold in a period of time) while, in nominal (or categorical) prediction, you are selecting what class a given observation belongs to.
Assess the performance of the trained model and record the Accuracy and Kappa scores for each parameter value the model used during training.
Use VarImp() to ascertain how the model prioritized each feature in the training (more on this in the resources)
TIP:
Remember that the data mining algorithms make the prediction in both types of tasks based on the similarities and differences between the attributes (columns) of observations (rows). You will be classifying "brand" in this task. 

Use Random Forest with 10-fold cross validation and manually tune 5 different mtry values. Using the same caret pipeline you developed for the previous step and the same training data, build a trained model using the Random Forest classifier in the caret package. After building the trained model
Assess the performance of the trained model and record the Accuracy and Kappa scores for each number of features the model used during training.
Use VarImp() to ascertain how the model prioritized each feature in the training (more on this in the resources)